+++ src/toisto/ui/text.py
@@ -97,8 +97,8 @@ class Feedback:
     def _try_again(self, guess: Label) -> str:
         """Return the feedback when the first attempt is incorrect."""
         if self.quiz.is_question(guess) and not self.quiz.has_quiz_type(GrammaticalQuizType):
-            colloquial = self.quiz.question.is_colloquial
-            try_again = self.TRY_AGAIN_IN_ANSWER_STANDARD_LANGUAGE if colloquial else self.TRY_AGAIN_IN_ANSWER_LANGUAGE
+            standard = self.quiz.question.is_colloquial and self.quiz.question.language == self.quiz.answer.language
+            try_again = self.TRY_AGAIN_IN_ANSWER_STANDARD_LANGUAGE if standard else self.TRY_AGAIN_IN_ANSWER_LANGUAGE
             return try_again % dict(language=ALL_LANGUAGES[self.quiz.answer.language])
         return self.TRY_AGAIN